__OTHER_FRAMEWORKS__/grafana/orbit-github-integration-main/requirements.txt
PyGithub==1.55
python-decouple==3.3
logfmt-logger==0.1.2 